projects/webrtc_signaling_server/requirements.txt
python-socketio>=5.10.0
aiohttp>=3.9.0
aiortc>=1.5.0
av>=10.0.0
opencv-python>=4.8.0